GIRA!!!

+++ Homes/src/main/java/codes/shiftmc/homes/particle/image/Offset.java
@@ -1,8 +1,8 @@
 package codes.shiftmc.homes.particle.image;
 
 public record Offset(
-        float x,
-        float y,
-        float z
+        double x,
+        double y,
+        double z
 ) {
 }
